Rewrote MPI to use more lightweight and portable custom struct type

+++ include/common.h
@@ -24,7 +24,7 @@ typedef struct
   double vy;
   double ax;
   double ay;
-  size_t partition;
+  unsigned int partition;
 } particle_t;
 
 //
